Begin implementation of sendfile func

+++ commons.h
@@ -30,7 +30,7 @@ FTP_COMMANDS parse_ftp_cmd(const char * client_message);
 
 int send_file_to_socket(const char * filename, int socket, DIRECTION direction);
 int recv_file_from_socket(const char * filename, int socket, DIRECTION direction);
-int get_file_size(FILE * file);
+long int get_file_size(FILE * file);
 
 
 #endif
